reviewed PR comments
